+++ src/components/CalendarPagination/CalendarPagination.jsx
@@ -1,9 +1,7 @@
 
 
-const CalendarPagination = () => {
+export const CalendarPagination = () => {
   return (
     <div>CalendarPagination</div>
   )
 }
-
-export default CalendarPagination;
